Prototypical implementation for basic duck typed methods
For now only the lang|equal? method on Int:Int is implemented.
Method registration is currently done from module_init
code which is stripped away by the linker (since it is not referenced
directly).  To avoid this we combine all bitcode files into one
single bitcode file, before putting it in the archive.  This forces
llvm to have all module_init functions from the various source files
in one list.